rt(d3d12): rename quad_render to draw_quad
